PanDerm/segmentation/requirements.txt
# Core deep learning framework
transformers==4.40.1

# Data science essentials
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1

# Image processing
opencv-python==4.9.0.80
pillow

# Utilities
requests==2.28.2
tqdm==4.65.2
PyYAML

tensorboard==2.16.2
lightning==2.2.1
wandb==0.16.4
imageio==2.34.0
scikit-image==0.22.0
segmentation-models-pytorch==0.3.3
MedPy==0.5.2
ftfy==6.2.0

